最热文章

+++ article/views.py
@@ -6,14 +6,27 @@ from django.http import HttpResponse
 from .models import ArticlePost
 from .forms import ArticlePostForm
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.core.paginator import Paginator
 
 def article_list(request):
-    articles = ArticlePost.objects.all()
-    context = {'articles':articles}
-    return render(request,'article/list.html', context)
+    if request.GET.get('order') == 'total_views':
+        article_list = ArticlePost.objects.all().order_by('-total_views')
+        order = 'total_views'
+    else:
+        article_list = ArticlePost.objects.all()
+        order = 'normal'
+    paginator = Paginator(article_list, 3)
+    page = request.GET.get('page')
+    articles = paginator.get_page(page)
+
+    context = {'articles': articles, 'order':order}
+    return render(request, 'article/list.html', context)
 
 def article_detail(request,id):
     article = ArticlePost.objects.get(id=id)
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
     article.body = markdown.markdown(article.body,
                                      extensions=[
                                          # 包含 缩写、表格等常用扩展
@@ -24,12 +37,13 @@ def article_detail(request,id):
     context = {'article': article}
     return render(request, 'article/detail.html', context)
 
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     if request.method == 'POST':
         article_post_form = ArticlePostForm(data=request.POST)
         if article_post_form.is_valid():
             new_article = article_post_form.save(commit=False)
-            new_article.author = User.objects.get(id=1)
+            new_article.author = User.objects.get(id=request.user.id)
             new_article.save()
             return redirect("article:article_list")
         else:
@@ -39,16 +53,22 @@ def article_create(request):
         context = {'article_post_form': article_post_form}
         return render(request, 'article/create.html', context)
 
+@login_required(login_url='/userprofile/login/')
 def article_delete(request, id):
     if request.method == 'POST':
         article = ArticlePost.objects.get(id=id)
+        if request.user != article.author:
+            return HttpResponse('抱歉，你无权删除此文章')
         article.delete()
         return redirect('article:article_list')
     else:
         return HttpResponse('只支持post请求')
 
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     article = ArticlePost.objects.get(id=id)
+    if request.user != article.author:
+        return HttpResponse('抱歉，你无权修改此文章')
     if request.method == 'POST':
         article_post_form = ArticlePostForm(data=request.POST)
         if article_post_form.is_valid():
